+++ heating_emissions/components/gridded_emissions_artifact.py
@@ -21,24 +21,44 @@ log = logging.getLogger(__name__)
 
 
 def build_gridded_artifact(
-    result: gpd.GeoDataFrame, resources: ComputationResources, output: str = 'co2', per_capita: bool = True
+    result: gpd.GeoDataFrame, resources: ComputationResources, output: str, per_capita: bool = True
 ) -> Artifact:
-    output_column = 'co2_emissions_per_capita'
-    file_name = 'heating_emissions_per_capita'
-    emission_type = 'Per capita'
-    legend_upper_cap = 3000
     legend_lower_cap = 0
-    tags = {Topics.EMISSIONS}
-    if not per_capita:
-        output_column = 'co2_emissions'
-        file_name = 'heating_emissions_absolute'
-        emission_type = 'Absolute'
-        legend_upper_cap = 150000
-        tags = {Topics.EMISSIONS}
+    if output == 'direct_co2_emissions':
+        output_column = 'direct_co2_emissions_per_capita'
+        file_name = 'direct_heating_emissions_per_capita'
+        emission_type = 'Per capita'
+        legend_upper_cap = 3000
+        legend_lower_cap = 0
+        tags = {Topics.DIRECT_EMISSIONS}
+        if not per_capita:
+            output_column = 'direct_co2_emissions'
+            file_name = 'direct_heating_emissions_absolute'
+            emission_type = 'Absolute'
+            legend_upper_cap = 150000
+            tags = {Topics.DIRECT_EMISSIONS}
 
-    layer_name = f'{emission_type} CO₂ emissions (kg/year)'
-    caption = f'{emission_type} CO₂ emissions from residential heating per year per 100-m pixel (Estimated)'
-    description = '**Estimated** territorial (scope 1) CO₂ emissions from heating residential buildings.'
+        layer_name = f'{emission_type} direct CO₂ emissions (kg/year)'
+        caption = f'{emission_type} direct CO₂ emissions from residential heating per year per 100-m pixel'
+        description = '**Estimated** direct CO₂ emissions from heating residential buildings.'
+
+    if output == 'life_cycle_co2_emissions':
+        output_column = 'life_cycle_co2_emissions_per_capita'
+        file_name = 'life_cycle_heating_emissions_per_capita'
+        emission_type = 'Per capita'
+        legend_upper_cap = 3000
+        legend_lower_cap = 0
+        tags = {Topics.LIFE_CYCLE_EMISSIONS}
+        if not per_capita:
+            output_column = 'life_cycle_co2_emissions'
+            file_name = 'life_cycle_heating_emissions_absolute'
+            emission_type = 'Absolute'
+            legend_upper_cap = 150000
+            tags = {Topics.LIFE_CYCLE_EMISSIONS}
+
+        layer_name = f'{emission_type} life cycle GHG emissions (kg CO₂eq/year)'
+        caption = f'{emission_type} life cycle GHG emissions from residential heating per 100-m pixel (kg CO₂eq/year)'
+        description = '**Estimated** life cycle GHG emissions from heating residential buildings.'
 
     if output == 'heat_consumption':
         output_column = output
@@ -65,16 +85,27 @@ def build_gridded_artifact(
         description = 'Average living space per capita (m²) in 100-m grid cells (data from 2022 German census)'
         tags = {Topics.PARAMETERS}
 
-    if output == 'emission_factor':
-        output_column = output
-        file_name = output
+    if output == 'direct_emission_factor':
+        output_column = 'direct_emission_factor'
+        file_name = 'direct_emission_factor'
         legend_upper_cap = 0.3
-        layer_name = 'Emission factor (kg of CO₂ per kWh)'
-        caption = 'Average (scope 1) emission factor from heating (Estimated)'
+        layer_name = 'Direct emission factor (kg of CO₂ per kWh)'
+        caption = 'Average direct emission factor from heating (Estimated)'
         description = (
-            '**Estimated** level of in situ (scope 1) emissions (kg of CO₂ per kWh) from heating residential buildings.'
+            '**Estimated** average direct emission factor (kg of CO₂ per kWh) from heating residential buildings.'
         )
-        tags = {Topics.PARAMETERS}
+        tags = {Topics.DIRECT_EMISSIONS}
+
+    if output == 'life_cycle_emission_factor':
+        output_column = 'life_cycle_emission_factor'
+        file_name = 'life_cycle_emission_factor'
+        legend_upper_cap = 0.3
+        layer_name = 'Life cycle emission factor (kg of CO₂eq per kWh)'
+        caption = 'Average life cycle emission factor from heating (Estimated)'
+        description = (
+            '**Estimated** average life cycle emission factor (kg of CO₂eq per kWh) from heating residential buildings.'
+        )
+        tags = {Topics.LIFE_CYCLE_EMISSIONS}
 
     if 'yearly_emissions' in output:
         output_column, output_year = output.split(':')
@@ -82,7 +113,7 @@ def build_gridded_artifact(
         layer_name = f'{emission_type} CO₂ emissions (Simulated, {output_year}) (kg/year)'
         caption = f'{emission_type} CO₂ emissions from residential heating per 100-m pixel (Simulated, {output_year})'
         description = (
-            f'Territorial (scope 1) CO₂ emissions from heating residential buildings (Simulated, {output_year}). '
+            f'Direct (scope 1) CO₂ emissions from heating residential buildings (Simulated, {output_year}). '
             f'This result is computed based on simulated heating demand by demand_ninja model.'
         )
         tags = {Topics.TEMPORAL}