+++ test/components/test_census_data.py
@@ -19,7 +19,7 @@ def test_get_census_tables_from_db(operator):
         db_connection=operator.ca_database_connection, raster_grid=raster_grid
     )
 
-    expected_columns = ['population', 'average_sqm_per_person', 'heat_consumption', 'emission_factor']
+    expected_columns = ['population', 'average_sqm_per_person', 'heat_consumption', 'direct', 'life_cycle']
     expected_columns_uncalc = ['dominant_age', 'dominant_energy']
 
     assert all([c in census_data.columns for c in expected_columns])