+++ test/components/test_gridded_emisions_artifacts.py
@@ -15,11 +15,14 @@ ROOT_DIR = Path(__file__).parent.parent.parent
 @pytest.mark.parametrize(
     'per_capita,output,expected_name',
     [
-        (True, 'co2', 'Per capita CO₂ emissions (kg/year)'),
-        (False, 'co2', 'Absolute CO₂ emissions (kg/year)'),
+        (True, 'direct_co2_emissions', 'Per capita direct CO₂ emissions (kg/year)'),
+        (False, 'direct_co2_emissions', 'Absolute direct CO₂ emissions (kg/year)'),
+        (True, 'life_cycle_co2_emissions', 'Per capita life cycle GHG emissions (kg CO₂eq/year)'),
+        (False, 'life_cycle_co2_emissions', 'Absolute life cycle GHG emissions (kg CO₂eq/year)'),
         (True, 'heat_consumption', 'Energy consumption (kWh/m²/year)'),
         (True, 'average_sqm_per_person', 'Living space (m² per person)'),
-        (True, 'emission_factor', 'Emission factor (kg of CO₂ per kWh)'),
+        (True, 'direct_emission_factor', 'Direct emission factor (kg of CO₂ per kWh)'),
+        (True, 'life_cycle_emission_factor', 'Life cycle emission factor (kg of CO₂eq per kWh)'),
     ],
 )
 def test_build_gridded_artifact(per_capita, output, expected_name, compute_resources):
@@ -27,11 +30,14 @@ def test_build_gridded_artifact(per_capita, output, expected_name, compute_resou
         {
             'x_mp_100m': [4500000, 4500100, 4500200],
             'y_mp_100m': [3500000, 3500100, 3500200],
-            'co2_emissions_per_capita': [100.0, 200.0, 250.0],
-            'co2_emissions': [1000.0, 2000.0, 2500.0],
+            'direct_co2_emissions_per_capita': [100.0, 200.0, 250.0],
+            'life_cycle_co2_emissions_per_capita': [300.0, 400.0, 450.0],
+            'direct_co2_emissions': [1000.0, 2000.0, 2500.0],
+            'life_cycle_co2_emissions': [2000.0, 4000.0, 5500.0],
             'heat_consumption': [80.0, 90.0, 100.0],
             'average_sqm_per_person': [20.0, 30.0, 40.0],
-            'emission_factor': [0.1, 0.2, 0.25],
+            'direct_emission_factor': [0.1, 0.2, 0.25],
+            'life_cycle_emission_factor': [0.4, 0.2, 0.25],
         }
     )