+++ test/components/test_utils.py
@@ -6,7 +6,7 @@ from shapely import box
 from heating_emissions.components.utils import (
     calculate_heating_emissions,
     get_aoi_area,
-    postprocess_uncalculate_census_data,
+    postprocess_uncalculated_census_data,
 )
 
 
@@ -16,7 +16,8 @@ def test_calculate_heating_emissions():
             'population': [31],
             'average_sqm_per_person': [40],
             'heat_consumption': [65],
-            'emission_factor': [0.2],
+            'direct': [0.2],
+            'life_cycle': [0.4],
             'latitude': [45.15],
             'longitude': [5.15],
         }
@@ -30,16 +31,26 @@ def test_calculate_heating_emissions():
 
     result = calculate_heating_emissions(census_data=example_geo_dataframe)
 
-    expected_per_capita = (
+    expected_per_capita_direct = (
         example_dataframe['average_sqm_per_person']
         * example_dataframe['heat_consumption']
-        * example_dataframe['emission_factor']
+        * example_dataframe['direct']
     )
 
-    expected_absolute = expected_per_capita * example_dataframe['population']
+    expected_absolute_direct = expected_per_capita_direct * example_dataframe['population']
 
-    assert result['co2_emissions'][0] == expected_absolute[0]
-    assert result['co2_emissions_per_capita'][0] == expected_per_capita[0]
+    expected_per_capita_life_cycle = (
+        example_dataframe['average_sqm_per_person']
+        * example_dataframe['heat_consumption']
+        * example_dataframe['life_cycle']
+    )
+
+    expected_absolute_life_cycle = expected_per_capita_life_cycle * example_dataframe['population']
+
+    assert result['direct_co2_emissions'][0] == expected_absolute_direct[0]
+    assert result['direct_co2_emissions_per_capita'][0] == expected_per_capita_direct[0]
+    assert result['life_cycle_co2_emissions'][0] == expected_absolute_life_cycle[0]
+    assert result['life_cycle_co2_emissions_per_capita'][0] == expected_per_capita_life_cycle[0]
 
 
 def test_calculate_heating_emissions_missing_data():
@@ -48,7 +59,8 @@ def test_calculate_heating_emissions_missing_data():
             'population': [31, 85, 15],
             'average_sqm_per_person': [np.nan, 80, 75],
             'heat_consumption': [65, 125.3, 145.2],
-            'emission_factor': [0.2, 0.15, 0.3],
+            'direct': [0.2, 0.15, 0.3],
+            'life_cycle': [0.4, 0.15, 0.3],
             'latitude': [45.15, 45.16, 45.15],
             'longitude': [5.15, 5.16, 5.16],
         }
@@ -62,9 +74,11 @@ def test_calculate_heating_emissions_missing_data():
 
     result = calculate_heating_emissions(census_data=example_geo_dataframe)
 
-    expected_absolute = round(31 * ((80 + 75) / 2) * 65 * 0.2)
+    expected_absolute_direct = round(31 * ((80 + 75) / 2) * 65 * 0.2)
+    expected_absolute_life_cycle = round(31 * ((80 + 75) / 2) * 65 * 0.4)
 
-    assert result['co2_emissions'][0] == expected_absolute
+    assert result['direct_co2_emissions'][0] == expected_absolute_direct
+    assert result['life_cycle_co2_emissions'][0] == expected_absolute_life_cycle
 
 
 def test_get_aoi_area():
@@ -81,6 +95,6 @@ def test_postprocess_uncalculate_census_data():
             'geometry': [None, None, None],
         }
     )
-    result = postprocess_uncalculate_census_data(df)
+    result = postprocess_uncalculated_census_data(df)
     assert all(result['dominant_age'] == ['Unknown', '1949-1978', 'Unknown'])
     assert all(result['dominant_energy'] == ['Unknown', 'Gas', 'Unknown'])